+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - EntityEditModel.setDefault() renamed defaultValue(), now Value based.
 - EntityEditModel.Insert, Update and Delete methods renamed.
 - EntityEditModel.Insert, Update and Delete split into separate steps, related refactoring.
+- EntityEditModel.refreshEntity() renamed refresh().
 
 ## 0.17.34
 ### is.codion.common.core